Happy Little Accident
Forgot to check some paths, and then the ModInfo for the list method ended up in the Links folder, which is actually a great idea and I'm glad this happened. Finalized now, so that should just work now going forward.
Also changed the window header to reflect the upcoming v1.3.0 update.